+++ src/types/cover_t.h
@@ -18,27 +18,27 @@ typedef struct cover_t
 	/**
 	 * Number of attributes
 	 */
-	uint32_t n_attributes;
+	uint64_t n_attributes;
 
 	/**
 	 * Total number of lines of the disjoint matrix
 	 */
-	uint32_t n_matrix_lines;
+	uint64_t n_matrix_lines;
 
 	/**
 	 * Number of words needed to store a line
 	 */
-	uint32_t n_words_in_a_line;
+	uint64_t n_words_in_a_line;
 
 	/**
 	 * Offset of the disjoint matrix
 	 */
-	uint32_t column_offset_words;
+	uint64_t column_offset_words;
 
 	/**
 	 * Number of words needed to store a column
 	 */
-	uint32_t column_n_words;
+	uint64_t column_n_words;
 
 	/**
 	 * Bit array of covered lines
@@ -48,7 +48,7 @@ typedef struct cover_t
 	/**
 	 * Number of lines that are not covered yet
 	 */
-	uint32_t n_uncovered_lines;
+	uint64_t n_uncovered_lines;
 
 	/**
 	 * Bit array of selected attributes
@@ -58,7 +58,7 @@ typedef struct cover_t
 	/**
 	 * Array with the current totals for all attributes
 	 */
-	uint32_t* attribute_totals;
+	uint64_t* attribute_totals;
 } cover_t;
 
 #endif // TYPES_COVER_T_H
